add all the viewmodel

+++ app/src/main/java/com/example/pollcreator/dataclass/Candidate.kt
@@ -1,7 +1,7 @@
 package com.example.pollcreator.dataclass
 
 data class Candidate(
-    val _pollId: Long,
+    val _pollId: Double,
     val _aadharNo: Long,
     val _age: Int,
     val _gender: Gender,
